Interactive Chart Scripts for all 3 shapes
I modified the BarScalers and have removed this line  "bar.transform.position = barPosition;" to prevent it from moving around. Added scripts to calculate S.A. & Volume respectively for both Cylinders and Pyramid. Multiplied the extortion of each component by a factor of 4 to make the shape move faster. Have tried to decrease scaling by a factor of 100, but it seems the volume and S.A. grow too fast, so it may be impossible to have it all scale proportionally (unless we want the S.A. and Volume to scale slower proportionally?) Also, I modified the included prefab to match the correct name (pyramid) & included the respective functioning scenes w/ a video of them working on Slack